+++ docs/changelog.md
@@ -11,6 +11,7 @@ PySceneDetect Releases
  * Primarily maintenance release including various bugfixes and minor API changes
  * Removed first row from statsfiles so it is a valid CSV file
  * The progress bar now correctly resizes when the terminal is resized
+ * Image sequences and URLs are now supported for input via the CLI/API
 
 #### Changelog
 
@@ -21,10 +22,12 @@ PySceneDetect Releases
  * [bugfix] Fix unhandled exception causing improper error message when a video fails to load on non-Windows platforms (resolves [#192](https://github.com/Breakthrough/PySceneDetect/issues/192))
  * [enhancement] Enabled dynamic resizing for progress bar (resolves [#193](https://github.com/Breakthrough/PySceneDetect/issues/193))
  * [enhancement] Always ouptut version number via logger to assist with debugging (resolves [#171](https://github.com/Breakthrough/PySceneDetect/issues/171))
+ * [enhancement] Allow image sequences and URLs to be used as inputs (resolves [#152](https://github.com/Breakthrough/PySceneDetect/issues/171) and [#188](https://github.com/Breakthrough/PySceneDetect/issues/188))
 
 #### Known Issues
 
  * Variable framerate videos are not supported properly currently (#168), a warning may be added in the next release to indicate when a VFR video is detected, until this can be properly resolved ([#168](https://github.com/Breakthrough/PySceneDetect/issues/168))
+ * Image sequences or URL inputs are not supported by the `save-images` or `split-video` commands
  * In certain cases, video files which will not load will fail silently, with PySceneDetect reporting that it processed 0 frames.  Better error handling/messaging is planned for these cases as part of [#179](https://github.com/Breakthrough/PySceneDetect/issues/179)